pyproject.toml
[project]
name = "routine-bot"
description = "Routine-tracking LINE bot"
version = "0"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cachetools>=6.2.1",
    "fastapi[standard]>=0.116.1",
    "line-bot-sdk>=3.18.1",
    "psycopg[binary]>=3.2.9",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]

[tool.pyright]
reportCallIssue = false
reportAttributeAccessIssue = false
reportOptionalMemberAccess = false
